[IMPORTANT] Mayor clean up
The plugin should load faster now (it doesn't import the python padlib
module anymore on initialization).

All configuration variables are namespaced under g:pad# now (instead of
g:pad_). For now, g:pad_dir is still used if present, but it is
recommended to update the configuration.

Implemented a new command API. The new :Pad command should be used
instead of :ListPads and :OpenPad (they are still available, but will be
   removed in the future).