Correctly generate the warning if vim-pad mappings conflict
I'm not sure how the code got this way, but it appears that previously the entire first half of the warning message was taken as the condition for the ternary that chooses between "normal" and "insert" for the message, meaning you always get the useless message "insert mode is already mapped", no matter which mode you're in. Very confusing!

This patch simply puts parentheses around the ternary expression so that the warning messages are generated as intended.